+++ app/src/main/AndroidManifest.xml
@@ -50,7 +50,7 @@
         <!--<activity android:name=".ui.events.EventsActivity"></activity>-->
         <activity android:name=".ui.event.EventActivity"></activity>
         <activity android:name=".ui.createevent.CreateEventActivity"></activity>
-        <activity android:name=".ui.login.LoginActivity" android:parentActivityName=".ui.main.MainActivity"></activity>
+        <activity android:name=".ui.login.LoginActivity" android:theme="@style/AppTheme" android:parentActivityName=".ui.main.MainActivity"></activity>
         <activity android:name="com.facebook.FacebookActivity"
                   android:configChanges="keyboard|keyboardHidden|screenLayout|screenSize|orientation"
                   android:label="@string/app_name"></activity>
